Adding context feature & fixing CI tests (#115)
Fixing race conditions, added context feature
