+++ Tests/TUSKitTests/TUSClient/TUSClient_DelegateTests.swift
@@ -1,0 +1,98 @@
+import XCTest
+import TUSKit // ⚠️ No testable import. Make sure we test the public api here, and not against internals. Please look at TUSClientInternalTests if you want a testable import version.
+final class TUSClient_DelegateTests: XCTestCase {
+    
+    var client: TUSClient!
+    var otherClient: TUSClient!
+    var tusDelegate: TUSMockDelegate!
+    var relativeStoragePath: URL!
+    var fullStoragePath: URL!
+    var data: Data!
+    
+    override func setUp() {
+        super.setUp()
+        
+        relativeStoragePath = URL(string: "TUSTEST")!
+        
+        MockURLProtocol.reset()
+        
+        let docDir = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0]
+        fullStoragePath = docDir.appendingPathComponent(relativeStoragePath.absoluteString)
+        
+        clearDirectory(dir: fullStoragePath)
+        
+        data = Data("abcdef".utf8)
+        
+        client = makeClient(storagePath: relativeStoragePath)
+        tusDelegate = TUSMockDelegate()
+        client.delegate = tusDelegate
+        do {
+            try client.reset()
+        } catch {
+            XCTFail("Could not reset \(error)")
+        }
+        
+        prepareNetworkForSuccesfulUploads(data: data)
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+        clearDirectory(dir: fullStoragePath)
+    }
+     
+    
+
+    // MARK: - Delegate start calls
+    
+    func testStartedUploadIsCalledOnceForLargeFile() throws {
+        let data = Fixtures.makeLargeData()
+        
+        try upload(data: data)
+        
+        XCTAssertEqual(1, tusDelegate.startedUploads.count, "Expected start to be only called once for a chunked upload")
+    }
+    
+    
+    func testStartedUploadIsCalledOnceForLargeFileWhenUploadFails() throws {
+        prepareNetworkForFailingUploads()
+        // Even when retrying, start should only be called once.
+        let data = Fixtures.makeLargeData()
+        
+        try upload(data: data, shouldSucceed: false)
+        
+        XCTAssertEqual(1, tusDelegate.startedUploads.count, "Expected start to be only called once for a chunked upload with errors")
+    }
+
+    // MARK: - Private helper methods for uploading
+
+    private func waitForUploadsToFinish(_ amount: Int = 1) {
+        let uploadExpectation = expectation(description: "Waiting for upload to finished")
+        uploadExpectation.expectedFulfillmentCount = amount
+        tusDelegate.finishUploadExpectation = uploadExpectation
+        waitForExpectations(timeout: 6, handler: nil)
+    }
+    
+    private func waitForUploadsToFail(_ amount: Int = 1) {
+        let uploadFailedExpectation = expectation(description: "Waiting for upload to fail")
+        uploadFailedExpectation.expectedFulfillmentCount = amount
+        tusDelegate.uploadFailedExpectation = uploadFailedExpectation
+        waitForExpectations(timeout: 6, handler: nil)
+    }
+    
+    /// Upload data, a certain amount of times, and wait for it to be done.
+    /// Can optionally prepare a failing upload too.
+    @discardableResult
+    private func upload(data: Data, amount: Int = 1, customHeaders: [String: String] = [:], shouldSucceed: Bool = true) throws -> [UUID] {
+        let ids = try (0..<amount).map { _ -> UUID in
+            return try client.upload(data: data, customHeaders: customHeaders)
+        }
+        
+        if shouldSucceed {
+            waitForUploadsToFinish(amount)
+        } else {
+            waitForUploadsToFail(amount)
+        }
+
+        return ids
+    }
+}
